Allow direct access on WebDAV for local access
The https and davs protocols for direct access are implemented the same
in ROOT. Both should be valid and davs is appropriate for file system
operations. While remote read is possible with either it is, at the
moment, much slower than xroot. Hence no changes were made regarding
direct access on the WAN.